third_party/blink/web_tests/platform/generic/http/tests/dom/document-contentType-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS iframes[0].contentDocument.contentType is "text/css"
PASS iframes[0].contentDocument.cloneNode(false).contentType is "text/css"
PASS iframes[1].contentDocument.contentType is "application/x-javascript"
PASS iframes[1].contentDocument.cloneNode(false).contentType is "application/x-javascript"
PASS iframes[2].contentDocument.contentType is "text/html"
PASS iframes[2].contentDocument.cloneNode(false).contentType is "text/html"
PASS iframes[3].contentDocument.contentType is "text/html"
PASS iframes[3].contentDocument.cloneNode(false).contentType is "text/html"
PASS iframes[4].contentDocument.contentType is "text/xml"
PASS iframes[4].contentDocument.cloneNode(false).contentType is "text/xml"
PASS iframes[5].contentDocument.contentType is "text/xml"
PASS iframes[5].contentDocument.cloneNode(false).contentType is "text/xml"
PASS iframes[6].contentDocument.contentType is "text/html"
PASS iframes[6].contentDocument.cloneNode(false).contentType is "text/html"
PASS iframes[7].contentDocument.contentType is "text/html"
PASS iframes[7].contentDocument.cloneNode(false).contentType is "text/html"
PASS iframes[8].contentDocument.contentType is "text/plain"
PASS iframes[8].contentDocument.cloneNode(false).contentType is "text/plain"
PASS iframes[9].contentDocument.contentType is "application/xml"
PASS iframes[9].contentDocument.cloneNode(false).contentType is "application/xml"
PASS iframes[10].contentDocument.contentType is "image/bmp"
PASS iframes[10].contentDocument.cloneNode(false).contentType is "image/bmp"
PASS iframes[11].contentDocument.contentType is "image/gif"
PASS iframes[11].contentDocument.cloneNode(false).contentType is "image/gif"
PASS iframes[12].contentDocument.contentType is "image/jpeg"
PASS iframes[12].contentDocument.cloneNode(false).contentType is "image/jpeg"
PASS iframes[13].contentDocument.contentType is "image/png"
PASS iframes[13].contentDocument.cloneNode(false).contentType is "image/png"

             
